LPD-25046 Rename package
